pyproject.toml
[project]
name = "notes"
version = "0.1.0"
description = "Self-hosted notes app with AI, calendar integrations, and full auth"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "litellm>=1.83.9",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "bcrypt>=4.0.0",
    "python-jose[cryptography]>=3.3.0",
    "cryptography>=42.0.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "fastapi-mail>=1.4.1",
    "google-auth>=2.29.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.130.0",
    "msal>=1.28.0",
    "uvicorn[standard]>=0.29.0",
    "aiofiles>=23.2.1",
    "pydantic-settings>=2.13.1",
    "itsdangerous>=2.2.0",
]

[dependency-groups]
dev = [
    "ruff>=0.4.0",
]
prod = [
    "pgvector>=0.3.0",
    "psycopg2-binary>=2.9.9",
]
